worker creates volume based on job setting
Openstack has the ability to launch a VM that creates it own Volume.
Here we add volume_size, boot_from_volume=True, terminate_volume=True to create_server.
This allocates a cinder volume of a particular size and uses the worker vm as the basis.
There is no need to format a volume and avoids several timing issues.
It also automatically cleans up the volume when the VM is deleted.

It expects an optional field called 'volume_size' in bespin-api job JSON.
This size is in GB and defaults to 100GB.