remove save output sidecar container
This approach had a problem that the job would run forever if the main
container failed. The sidecar process would just keep waiting for a file that never
exists. Will create an issue to come up with a better approach.
For now the project_id and readme_file_id will be "TODO".